Fix json serialization error in random forest scoring test

+++ regression-tests/sparktkregtests/testcases/scoretests/random_forest_test.py
@@ -29,7 +29,7 @@ class RandomForest(sparktk_test.SparkTKTestCase):
         """Build the required frame"""
         super(RandomForest, self).setUp()
 
-        schema = [("feat1", int), ("feat2", int), ("class", int)]
+        schema = [("feat1", float), ("feat2", float), ("class", int)]
         filename = self.get_file("rand_forest_class.csv")
 
         self.frame = self.context.frame.import_csv(filename, schema=schema)
